+++ examples/counter/web-remix/app/entry.client.tsx
@@ -5,18 +5,13 @@
  */
 
 import { RemixBrowser } from "@remix-run/react";
-import { startTransition, StrictMode } from "react";
+import { startTransition } from "react";
 import { hydrateRoot } from "react-dom/client";
 import init from "shared/shared";
 import wasm from "shared/shared_bg.wasm";
 
 init(wasm).then(() => {
   startTransition(() => {
-    hydrateRoot(
-      document,
-      <StrictMode>
-        <RemixBrowser />
-      </StrictMode>
-    );
+    hydrateRoot(document, <RemixBrowser />);
   });
 });
