+++ examples/counter/web-remix/app/routes/_index.tsx
@@ -1,32 +1,13 @@
 import type { V2_MetaFunction } from "@remix-run/node";
 import { useEffect, useState } from "react";
 
-import init_core, { process_event, handle_response, view } from "shared/shared";
-import type {
-  Event,
-  HttpResponse,
-  SseResponseVariantChunk,
-  SseResponseVariantDone,
-} from "shared_types/types/shared_types";
 import {
-  EffectVariantRender,
   ViewModel,
-  EffectVariantHttp,
-  EffectVariantServerSentEvents,
   EventVariantStartWatch,
   EventVariantDecrement,
   EventVariantIncrement,
-  Request,
 } from "shared_types/types/shared_types";
-import {
-  BincodeSerializer,
-  BincodeDeserializer,
-} from "shared_types/bincode/mod";
-
-import { request as http } from "../http";
-import { request as sse } from "../sse";
-
-type Response = HttpResponse | SseResponseVariantChunk | SseResponseVariantDone;
+import { update } from "../core";
 
 export const meta: V2_MetaFunction = () => {
   return [
@@ -37,60 +18,10 @@ export const meta: V2_MetaFunction = () => {
 
 export default function Index() {
   const [state, setState] = useState(new ViewModel("", false));
-
-  function dispatch(event: Event) {
-    const serializer = new BincodeSerializer();
-    event.serialize(serializer);
-    const effects = process_event(serializer.getBytes());
-    processEffects(effects);
-  }
-
-  function respond(uuid: number[], response: Response) {
-    const serializer = new BincodeSerializer();
-    response.serialize(serializer);
-    const effects = handle_response(
-      new Uint8Array(uuid),
-      serializer.getBytes()
-    );
-    processEffects(effects);
-  }
-
-  async function processEffects(effects: Uint8Array) {
-    const requests = deserializeRequests(effects);
-
-    for (const { uuid, effect } of requests) {
-      switch (effect.constructor) {
-        case EffectVariantRender: {
-          setState(deserializeView(view()));
-          break;
-        }
-        case EffectVariantHttp: {
-          const request = (effect as EffectVariantHttp).value;
-          const response = await http(request);
-          respond(uuid, response);
-          break;
-        }
-        case EffectVariantServerSentEvents: {
-          const request = (effect as EffectVariantServerSentEvents).value;
-          for await (const response of sse(request)) {
-            respond(uuid, response);
-          }
-          break;
-        }
-      }
-    }
-  }
-
   useEffect(
     () => {
-      async function loadCore() {
-        await init_core();
-
-        // Initial event
-        dispatch(new EventVariantStartWatch());
-      }
-
-      loadCore();
+      // Initial event, beware of StrictMode in ../entry.client.tsx as it will run twice in dev
+      update(new EventVariantStartWatch(), setState);
     },
     // eslint-disable-next-line react-hooks/exhaustive-deps
     /*once*/ []
@@ -107,13 +38,13 @@ export default function Index() {
         <div className="buttons section is-centered">
           <button
             className="button is-primary is-warning"
-            onClick={() => dispatch(new EventVariantDecrement())}
+            onClick={() => update(new EventVariantDecrement(), setState)}
           >
             {"Decrement"}
           </button>
           <button
             className="button is-primary is-danger"
-            onClick={() => dispatch(new EventVariantIncrement())}
+            onClick={() => update(new EventVariantIncrement(), setState)}
           >
             {"Increment"}
           </button>
@@ -121,19 +52,4 @@ export default function Index() {
       </section>
     </main>
   );
-}
-
-function deserializeRequests(bytes: Uint8Array) {
-  const deserializer = new BincodeDeserializer(bytes);
-  const len = deserializer.deserializeLen();
-  const requests: Request[] = [];
-  for (let i = 0; i < len; i++) {
-    const request = Request.deserialize(deserializer);
-    requests.push(request);
-  }
-  return requests;
-}
-
-function deserializeView(bytes: Uint8Array) {
-  return ViewModel.deserialize(new BincodeDeserializer(bytes));
 }
