+++ examples/simple_counter/web-remix/app/routes/_index.tsx
@@ -1,20 +1,13 @@
 import type { V2_MetaFunction } from "@remix-run/node";
 import { useEffect, useState } from "react";
 
-import { process_event, view } from "shared/shared";
-import type { Event } from "shared_types/types/shared_types";
 import {
-  Request,
   ViewModel,
-  EffectVariantRender,
   EventVariantReset,
   EventVariantIncrement,
   EventVariantDecrement,
 } from "shared_types/types/shared_types";
-import {
-  BincodeDeserializer,
-  BincodeSerializer,
-} from "shared_types/bincode/mod";
+import { update } from "../core";
 
 export const meta: V2_MetaFunction = () => {
   return [
@@ -26,30 +19,10 @@ export const meta: V2_MetaFunction = () => {
 export default function Index() {
   const [state, setState] = useState(new ViewModel("0"));
 
-  function dispatch(event: Event) {
-    const serializer = new BincodeSerializer();
-    event.serialize(serializer);
-    const effects = process_event(serializer.getBytes());
-    processEffects(effects);
-  }
-
-  async function processEffects(effects: Uint8Array) {
-    const requests = deserializeRequests(effects);
-
-    for (const { effect } of requests) {
-      switch (effect.constructor) {
-        case EffectVariantRender: {
-          setState(deserializeView(view()));
-          break;
-        }
-      }
-    }
-  }
-
   useEffect(
     () => {
       // Initial event
-      dispatch(new EventVariantReset());
+      update(new EventVariantReset(), setState);
     },
     // eslint-disable-next-line react-hooks/exhaustive-deps
     /*once*/ []
@@ -62,19 +35,19 @@ export default function Index() {
         <div className="buttons section is-centered">
           <button
             className="button is-primary is-danger"
-            onClick={() => dispatch(new EventVariantReset())}
+            onClick={() => update(new EventVariantReset(), setState)}
           >
             {"Reset"}
           </button>
           <button
             className="button is-primary is-success"
-            onClick={() => dispatch(new EventVariantIncrement())}
+            onClick={() => update(new EventVariantIncrement(), setState)}
           >
             {"Increment"}
           </button>
           <button
             className="button is-primary is-warning"
-            onClick={() => dispatch(new EventVariantDecrement())}
+            onClick={() => update(new EventVariantDecrement(), setState)}
           >
             {"Decrement"}
           </button>
@@ -82,19 +55,4 @@ export default function Index() {
       </section>
     </main>
   );
-}
-
-function deserializeRequests(bytes: Uint8Array) {
-  const deserializer = new BincodeDeserializer(bytes);
-  const len = deserializer.deserializeLen();
-  const requests: Request[] = [];
-  for (let i = 0; i < len; i++) {
-    const request = Request.deserialize(deserializer);
-    requests.push(request);
-  }
-  return requests;
-}
-
-function deserializeView(bytes: Uint8Array) {
-  return ViewModel.deserialize(new BincodeDeserializer(bytes));
 }
